Always start form jupyter/notebook
By removing the dependence of the Docker Containers from each other and only depend on `jupyter/notebook` the download time is increased. At the same time the recipe to build the container is drastically simplified as it is the same recipe for all containers. This would allow us to build the containers directly in the corresponding repositories.